attr: Shrink slices with cap » len
If there are a lot of players - as when testing networking - the backing array
for the inventory content slices can balloon and not be reclaimed. This change
'shrinks' the slices when elements are removed if:

  cap - (len * f) > b

Where cap is the slice's capacity, len is it's length, f is the factor of how
many times larger the length can be compared to capacity, b is a buffer of the
minimum gain when shrinking the slice.

Assuming f=2 - capacity can be upto double the length - and a buffer of 4 -
when shrinking we want a difference or gain of at least 4 elements worth of
storage to be reclaimed as there is no point wasting cycles shrinking if there
is little to be gain.

For a slice with length 10, capacity 20 and a buffer of 4 would not be
considered shrinkable:

  20 - (10*2) > 4
  20 - 20 > 4
  0 > 4
  false

If the capacity was increased to 22 it would still not be considered
shrinkable because the gain would be small, less than b:

  22 - (10*2) > 4
  22 - 20 > 4
  2 > 4
  flase

If the capacity was increased to 25 it would be considered shrinkable as the
capacity is more than twice the length and the gain is more than 4 elements:

  25 - (10*2) > 4
  25 - 20 > 4
  5 > 4
  true, so reclaimed

The buffer prevents us shrinking slices where there is little gain.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>