cmd: Stash actor's location in state
We need to look up the actor's location so that we can add the lock. We might
as well stash this in the state are most commands also need to lookup the
actor's location so this will save some cycles. Commands will need updating to
take advantage of the stashed value.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>