cmd: Add crowd control to LOOK command
This change adds 'crowd control' to the look command. If there are more than
crowdSize players at a location instead of listing all of the players it will
list "You see a crowd here".

As a side effect this also puts the players before objects in the list of
things at a location.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>