cmd: Improved version of LOOK command
The LOOK command is a major bottleneck to supporting a large number of
players. This change switches the LOOK command to use the stashed state.where
as well as implementing an 'optimised' list of things at the location if it is
crowded. Actually it just says "You see a crowd here." instead of iterating
over all the items and mobiles and listing them.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>