comms: Redraw prompt when just hitting enter
If a user just hits enter with no input the cursor just moved down a line in
the client. In this instance the prompt should be redrawn.

Fixing this issue removes the line feed from the defaultPrompt. As the prompt
is drawn on it's own line we can actually add it to the output - with a
network carriage return + linefeed - after folding.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>