config,stats: Add knob for running GC before stats
Added a Stats.GC to configure wether a GC and FreeOSMemory should be done
before stats are collected. This can help to keep resources low. If set to
true it's probably a good idea not to collect stats too often. If set to false
the runtime will do it's normal GC and scavenging.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>