attr: Update FindName to return typed nil + name preset
As well as updating FindName to return a *Name typed nil this change also adds
a name preset to the Name method. This acts as a default name if there is no
attribute or if the attribute has a unset name. I decided on 'preset' as
'default' cannot be used due to it being a reserved keyword :|

The preset was added at the same time as the typed nil as it fitted well with
the required changes. For example:

  who := "Someone"
  if a := attr.FindName(s.actor); a != nil {
    who = a.Name()
  }

Becomes:

  who := attr.FindName(s.actor).Name("someone")

It also works well with a common pattern of using the player's word as the
initial initial value when parsing commands:

  name := s.words[0]

  // Other interesting code...

  name = attr.FindName(what).Name(name)

Here the initial s.words[0] is used if there is no name available.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>