attr,cmd: Reimplement narratives
Original plan was to have a Narrative 'inventory' which was a special type of
Inventory. The idea was that narrative items were less common than
non-narrative items and so processing could be reduced most of the time by
ignoring narratives. However in implementing commands we were always checking
the main Inventory for non-narrative items and then Narratives for narrative
items which over complicated things and added code to command processing.

The *only* difference between narrative and non-narrative items are:

  - They are not listed in location/container descriptions
  - You cannot pick them up at all

The Narrative attribute now only flags or 'tags' an item as being a narrative.
Narrative and non-narrative items are all treated the same and all contained
with a normal Inventory.

The only special processing for narratives are:

  - The GET command won't let you get narratives
  - Inventories will not list narratives

All other processing for narratives has been removed :)

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>