cmd,comms: Unify sending messages via messenger
This change unifies the sending of buffered messages via the state.messenger
method. Before only messages to the participant and observers were sent via
the messenger and messages to the actor were filtered back to the comms client
type through the state.Parse function.

Now all messages are sent using the state.messenger method and message text
and ok status are no longer fed back to the client type.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>