has: Split common.go into individual files
The has package has had a single file defining interfaces for ages. It has
always been the intention to separate the interfaces out into their own files.
By separating the out we can commit changes to specific interfaces and writing
tests for them will be a lot easier and cleaner in the future.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>