attr,has: Move Found into Attribute interface
Currently only a few attributes implement the Found method. Found is now part
of the Attribute interface and must be implemented by all attributes. There is
no default Found implementation and it must be implemented by each attribute.
This is due to the face that Found needs to use the receiver type of the
attribute so that we can check for an interface being nil when the type is not
nil - as opposed to checking for a simple nil.

For more details see the Go FAQ: "Why is my nil error value not equal

  https://golang.org/doc/faq#nil_error

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>