comms: Fix blank line when just hitting enter
Recent work to unify messaging has introduced an empty, blank line when the
player is just hitting enter. This change just redisplays the prompt and has
the side effect of bypassing the unnecessary call to Fold when the output is
empty.

We still want to display the prompt as later on it will show play information
and just hitting enter will still allow the information in the prompt to be
updated.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>