config,server,attr: Reimplement config using recordjar
Config now uses the default configuration file config.wrj. Inventory has been
updated to use a single configuration setting config.Inventory.Compact instead
of config.ReclaimBuffer and config.ReclaimFactor. If the capacity of an
inventory is more than twice its length and the difference is more than
config.Inventory.Compact the inventory will be compacted.

NOTE: The server can be run in one of two ways. You can change to the
directory where the configuration file is. Or you can specify where the config
file is on the command line. For example:

  cd /home/diddymus/gits/WolfMUD/bin
  ./server ../src/code.wolfmud.org/WolfMUD.git/data

OR:

  cd /home/diddymus/gits/WolfMUD/src/code.wolfmud.org/WolfMUD.git/data
  ../../../../bin/server

OR:

  cd /home/diddymus/gits/WolfMUD/src/code.wolfmud.org/WolfMUD.git/data
  /home/diddymus/gits/WolfMUD/bin/server

OR:

  /home/diddymus/gits/WolfMUD/bin/server /home/diddymus/gits/WolfMUD/src/code.wolfmud.org/WolfMUD.git/data/

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>