comms: Don't setup world during program initialisation
Setup requires configuration values which are lot loaded at initialisation
time. Also we are using init methods in Attribute derived types for registry.
The actual call to Setup is only where it is at the moment for convenience so
moving it from initialisation time to when the listener starts is not a big
thing.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>