stats: Always recreate list slice on last logout
When the last player logs out we now always recreate the player list slice
without checking the slice's capacity. The capacity check was arbitrary anyway
and just recreating an empty slice seems to help the garbage collector when
there have been a lot of players. Even if it does not help the GC it tidies up
the code a bit.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>