cmd: Delay allocation of state message buffers
Currently message buffers (actor, participant & observers) are allocated when
a new state is created and before locks are taken (BRL - big room locks). This
means if there are a number of players waiting for the same locks each will
have allocated buffers that are not used.

This change allocates message buffers once locks have been acquired.

To keep with pre-change functionality the buffers are not deallocated if it is
necessary to release and reacquire the locks. This helps to reduce allocation
churn. However this may be changed in the future.

To methods have been added allocateBuffers and deallocateBuffers to centralise
allocation and deallocation of the buffers for a state instance.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>