comms: Force quit player if still in the world
Stop checking for the 'quitting' error to determine if a player should be
forced to quit or if they did it voluntarily. The quitting error is to be
removed shortly but we can now use the new stats.Find function to determine if
the player needs to force quit to remove them from the world or if they did it
voluntarily. Force quits are usually used to clean up after a terminated
network connection due to an error.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>