cmd: Drop dispatcher and stop passing it around
At some point in development I must have thought passing the dispatcher around
as a func(*state) was a really good idea. Can't remember why now though.

Now we can simplify things if we stop passing the dispatcher around and drop
the dispatch function entirely. The handler lookup and invocation that
dispatcher was doing can be moved into sync.

We can also remove the duplicated effort of checking the length of the locks
done by both parse and sync. Instead the check is performed by sync which
returns a bool: true if the list of locks does not change (in sync) false
otherwise (not in sync).

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>