cmd: Remove redundant nil check
We don't need to check for nil as AddLock does the checking for itself anyway.
For the vast majority of the time s.where will not be nil. For the few cases
where it is nil we spend a few extra cycles calling AddLock which just returns
immediately. Spending a few extra cycles is worth it for the cleaner code.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>