frontend: Initial driver - work in progress
This is a very early version of a frontend driver for WolfMUD. It contains an
'account' front end that asks for the name of a player's character. It then
passes over control to a 'game' driver. The two drivers need to be split in to
their own source files but for now this will allow additional cleanup of the
client - comms/client.go - source. Note that these drivers are not complete
and this is very much a work in progress commit.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>