comms: Switch client to use new frontend package
This change drops all of the player creation, player Id generation, player
removal from the game and any other player related code. It also switches from
handling the network data as strings to handling them as bytes. All of this is
replaced by the new frontend package. All the client does now is read and
write network data and monitor for errors.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>