comms: Move prompt code out of client
We need two different prompts. The first if for the frontend menu prompts. The
second is a dynamic prompt when in the game. The first prompt is static and
now handled by the frontend drivers. The second prompt is now handled by
attr/player. It is currently static but will eventually be able to show play
stats like health, mana, etc. The client would not be able to handle the
second prompt without being able to reach into the game code, besides client
is probably not the right place for prompt handling anyway - it should just
handle network I/O and network error handling.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>