frontend: Login using old player files (WIP)
This commit allows players to login using existing player files from the
previous version of WolfMUD before the re-re-write. It is a work in progress
and needs a lot of cleaning up and refactoring. Account creation is NOT done
yet so for this commit you can only login with an existing player file :P

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>