frontend: Refactor EndOfDataError as driverClosedError
Due to recent work on the frontend package the function of the terribly name
EndOfDataError has changed. To reflect it's proper purposed more accuratly it
has been renamed to driverClosedError and is no longer exported. All of the
code dealing with the driverClosedError is now only found in the main
driver.go source file. The driverClosedError and associated functions also
have improved comments.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>