frontend: drop confusing Driver.write field
At the moment writing output back to a comms/client is done by both the
frontend.Driver and cmd.state. When we are not in the game only the Driver is
writing output. When in game both the Driver and state write output causing
weird behaviour like duplicate prompts.

The Driver.write field was intended as a flag to the Driver as to whether it
should be writing output or not. However 'write' by itself was confusing.

In an effort to clean up the code and misunderstandings the Driver.write field
has been dropped. Instead we simply take away the Driver's buffer, set
Driver.buf to nil, to stop it from writing. We give the buffer back, create a
new buffer, when we want the Driver to write output again.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>