frontend: Rename exported Driver→unexported frontend)
Part 2/2 of renaming exported Driver to unexported frontend.

Originally the Driver was seen as a way to drive I/O between the frontend
menus and the backend game. Further development has made the term Driver less
meaningful. It has also meant that the Driver no longer needs to be exported.

This commit therefore renames the Driver type to just frontend, which is no
longer exported. All references have been updated to reflect this. There is
only a single place outside of the frontend package, in comms/client.go, where
the exported frontend.Driver was used. This is now taken care of via an
interface in the client type.

As there is no longer a Driver the driverClosedError is now just closedError.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>