frontend: Refactor sub frontends, drop greeting.go
The sub frontends - such as login and account - now have their own
types. This allows for splitting up the large number of methods on
frontend into the different smaller sub frontends. This will also allow
for data to be associated with the appropriate sub frontend instead of
using the frontend stash as a place to dump data. Eventually the
frontend stash will be dropped totally.

The file greeting.go has been dropped and the displayGreeting method has
been moved back into the main frontend. This is due to the fact that the
frontend struct nextFunc field needs to be initialised with an initial
func() type and the greeting seems to be a logical choice.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>