config: Default to data directory in current directory
If no directory is given on the command line the server will try to use
the 'data' directory in the current working directory by default. This
means that Windows users using a binary distribution will be able to
double-click the server to launch it. This may also help some Linux
desktop setups but it has not been tested. It does mean Linux users
using the binary distributions can open a terminal, cd to the server
directory and just type './server' to start a server.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>