comms: Stop fiddling with the client's err chan
Before the client Write method ignored temporary errors it was necessary
to clear the temporary errors before sending more data to the client,
such as saying goodbye or informing that the connection had timed out.
Now that the client Write method checks for temporary errors we can stop
fiddling with the err chan.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>