comms: Prevent misuse of err instead of c.Error()
Corrected code that was using err as a fallthrough side effect of the
main processing loop exiting. The err var was only intended for use in
the main processing loop. Variables used repeatedly by the loop, but
defined outside of it only once, are now scoped to the loop by a set of
outer braces. This means they cannot be misused again. Where err was
being misused replace with calls to c.Error().

By using err instead of c.Error() we may have missed errors not set in
the main processing loop. For example if another client goroutine
performs a Write and there is a connection issue. In this case the error
would have been flagged and the broken client shutdown correctly but we
wouldn't know why.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>