config: Switch config to using an init function
So that the configuration values are setup and immediatly available to
other packages the configuration package now uses an init function
instead of specifically calling load. Without using an init function
only the default values, before loading a configuration file, are
available which is not the intent.

The setting of the default logging format and initialisation of the
global random seed has also been moved from the server package into the
configuration package so that they are also immediatly set up.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>