comms: Cleanup and fix Accept error handling
Currently if there is an error accepting an incoming network connection
the Listen function will return causing the server to terminate. This
issue is easy to overlook as it is wrapped in a clumsy if/else block.

This change simplifies the logic inside the main for loop. The error
checking now logs any errors and then continues with the loop and
accepting the next connection.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>