comms: Recover panicking clients, isolate cleanup code
At the moment a panicking client will bring down the whole server. This
change isolates the client cleanup code in a new close method. The close
method can then be called in a defered function. In the defered function
we first try to recover a panicking client - if not in debug mode -
before trying to close down the client goroutine gracefully.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>