cmd: add suppressible messaging to state script method
Enhanced the script method so that messages to the actor, participant
and observers can be suppressed via parameters. This is so that we can
have a single, unified way of scripting.

For common flag configurations there are some helper methods: scriptAll,
scriptNone and scriptActor.

The quit command has been updated to use scriptAll.

The script method can now be used in place of the silent method - which
can now be removed.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>