cmd: Make imput member of state behave as intended
It was originally intended that state.input would represent the input as
originally entered by a player. This is important for commands such as
SAY and TALK where we need the sanitized original text.

When state.input was sanitized the sanitized version was stored in
state.words and there was a one-to-one mapping of elements between
state.input and state.words.

When stop-word removal was introduced the stop-words were removed from
state.input and state.words losing the original input text. This was
noted at the time as being a bug.

This change removes the experimental status of stop-words. It also only
removes stop-words from state.words and leaves state.input as the
untouched original. This does mean there is no longer a one-to-one
mapping between state.input and state.words elements. However this does
not seem to be an issue as the mapping was never relied upon in any
code.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>