text: Handle widths less than 1 in Fold
If a width less than -31 is passed to Fold it will panic when creating
the line buffer. For Fold it does not make sense to have a width passed
that is less than 1. For widths less than 1 this change causes text not
to be folded but line endings are still changed from Unix '\n' to
network '\r\n'.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>