text: Suppress line feed when word length >= width
When the length of a single word matches or exceeds the folding width
and is followed by a line feed we end up with two line feeds instead of
one. This is because we add a line feed after reaching the folding width
and then add the natural line feed from the input text as well.

This change suppresses the added line feed when a single word matches or
exceeds the folding width so that only the natural line feed from the
input is added.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>