cmd/internal: Unexport Buffer and Buffers types
Now we have an exportd Msg struct and the Allocate/Deallocate methods
are in buffer.go instead of state.go there is no longer a need to have
either the Buffer or Buffers exported. This should result in being able
to update/change the buffer implementation as needed without impacting
external code.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>