cmd: Standardise on WriteStrings before method renamed
As part of the buffer work there will be a single Send method to send
messages instead of the current Write, WriteString, WriteStrings,
WriteByte, WriteRune methods. Updating references from WriteString to
WriteStrings will ease the transition. This change only contains
instances where the rename is just WriteString to WriteStrings. Changes
beyond a simple rename will be in a different commit.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>