cmd: Rename WriteStrings/WriteAppend→Send/Append
Initially WriteStrings - and previously WriteJoin - were named to fit in
with the other byte.Buffer write methods: WriteByte, WriteRune and
WriteString. Now we have migrated from bytes.Buffer to a []byte.

When a player interacts with WolfMUD using commands we send them
messages informing them of what happened. We also send players messages
when something happens near them. These message are sent via buffers. It
now makes sense to reflect this by renaming WriteStrings to Send and
WriteAppend to just Append. Now we can Send a message and Append to a
message which reads a lot better in code.

Rewriting of sources in cmd done mechanically using:

  gofmt -w -r 'a.WriteStrings->a.Send' cmd/
  gofmt -w -r 'a.WriteAppend->a.Append' cmd/

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>