test: Treat '␠' U+2420 as a hard space in Fold
It is sometimes desirable to have hard spaces in text that are not
foldable and not reducible to a single space. For this purpose the
Unicode character '␠' U+2420 can be used which will be turned into a
space on output.

This is a really, really ugly hack to make embedded spaces work without
rewriting Fold using a different algorithm just now :(

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>