config: Add NONE runtime configuration file
When writing tests for the server it is useful to be able to use a basic
configuration setup with no zones loaded. This is because some files
like attr/inventory.go use the configuration values in their init
functions. To accommodate this, it is now possible to pass NONE as the
configuration file name on the command line or via go test -args. When
NONE is used the server will use the built in default configuration
values only and no data files will be loaded.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>