cmd: Check delivery is to player in state.messenger
When calling attr.FindPlayer we could get a *Player typed nil returned.
In which case Deliver is calling Player.Write which just bails. Might as
well check what we find and avoid calling Deliver and Player.Write if we
can. This is especially important for Observers otherwise we end up trying
to deliver messages to non-players (i.e. items) in a location as well as
to players.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>