cmd/internal: Fix prefix, count & comments for buffer.Append
When appending messages to a buffer a space prefix should only be used
when the buffer isn't empty and the last character in the buffer is not
a line feed.

If the buffer is empty (we have not had a Send yet) we do not want the
leading space, otherwise it will cause the initial message to be
indented.

If the last character in the buffer is a line feed we do not want to add
a space. If we do add a space it will cause the message to be indented.

If Append is called without a prior call to Send we need to increase the
message count by one when appending an initial message.

Comments clarified and special cases added.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>