cmd/internal: Reimplement buffers.Silent, update cmd.state
The current buffers.Silent method is broken because you cannot map the
returned silent flags back to the corresponding buffer. This is because
the sequence of flags returned is undefined because the sequence of
ranging over a map is undefined. The new implementation returns two
buffers, one where all the flags were originally true and the other
where the flags were originally false. Returning buffers makes sense as
we can now use other buffers methods with them easily.

Also updated cmd.state to use new version of silent. This is a good
example of setting the silent flag for a buffers and then resetting the
flags back to their original state.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>