cmd/internal: buffers.Len returns map[has.Inventory]int
The buffers.Len method has been updated to return map[has.Inventory]int.
This is so that the message counts can be mapped back to a specific
buffer. Before the sequence of the []int was used. However the sequence
of []int is indeterminable because when ranging over a map the sequence
is indeterminable.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>