message: Use count not len() in buffer (and tests)
Should be checking the message count for a buffer instead of using
len(b.buf) to see if the buffer is empty or not. Except for one specific
case where we actually want to use len(b.buf) to find the last character
in the buffer. Using the count allows for leading line feeds and other
non-message data to be put into the buffer without effecting the message
count. It also means that we avoid the is "empty really empty" issues we
had previously.

Updated buffer tests to also use count instead of len().

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>