frontend: Add Write method to frontend
When in game messages go through the following pipeline:

  message.buffer.Deliver->attr.player.Write->client.Write

It is attr.player.Write that does any prompt processing. By adding Write
to the frontend we can have a similar setup:

  message.buffer.Deliver->attr.player.Write->client.Write

This will simplify maintenance and means we don't have two different
ways of doing the same thing.

NOTE: Write is not actually in use just yet as additional buffer work is
needed.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>