message: Deliver handles prompt LF, takes multiple writers
The buffer.Deliver has been updated to take multiple Writers to deliver
messages to. This is so that it can handle the buffers and make sure
that copies are sent to receivers except for when there is a single
receiver. The state.messenger method has been updated to use multiple
io.Writers for Observers.

buffer.Deliver now also handles the appending of the line feed required
to make sure a prompt is displayed on its own new line. Having deliver
handle it in one place has a few advantages: attr.player and frontend
can be simplified by dropping the processing, the buffer can be modified
once before it's sent to all receivers and finally it is handled
consistently. The tests have been updated to expect the new line feed.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>