cmd: state.tokenizeInput panics with only stopwords
If a player only enters stopwords at the prompt such as 'from' this
reduces the input to an empty string in tokenizeInput. It then causes
tokenizeInput to panic causing the client connection to be closed down
gracefully.

This fix changes the command to "Eh?" if removing stopwords results in
an empty string. The parser does not understand "Eh?" as being a valid
command and so responds with "Eh?" to the player as if they had entered
an invalid command in the first palace.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>