attr: Implement Vetoes.Unmarshal fully and document
Vetoes can now be unmarshalled from recordjar files. Vetoes are
specified using the new Keyed String List data type. Vetoes are a way of
preventing an action on something that would normally be allowed.

For example:

  VETOES: GET→You cannot get the magical stone.
        : EXAMINE→You try to examine the magical stone but doing so
          makes your eyes water and sting.

This would veto the commands GET and EXAMINE giving the specified reason
for each - presumably as part of the definition of a 'magical stone'.

Documentation for zone files now has an entry for vetoes.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>