cmd: Amalgamate parse.go into handler.go and state.go
The file parse.go was the original parser. Over time it became just an
entry point into the parsing and processing routines in the state.go
file. This separation of parse and state has become obsolete to the
point where the file can now be removed and it's remaining processing
pushed into the handler.go and state.go files.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>