cmd: Add state.scripting field and Script function
The state.scripting field controls whether the scripting only commands -
that start with a '$' symbol - are executable. For the Parse function
state.scripting is false. For the new Script function state.scripting is
true. For the method state.script state.scripting is set to true and
then restored to its previous value.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>