attr: Always call Locate.SetWhere in Inventory.Add
Currently adding items into an Inventory via Inventory.Add only updates
the Locate attribute for non-narrative items. This was for performance
reasons as it avoided calling FindLocate for narratives - which tend to
be non-moveable. However a narrative may need to know where it is to
interact with its environment. For narrative items the Locate attribute
is usually updated only once - when initially added to the Inventory -
so the performance impact is actually negligible. This change causes the
Locate attribute of all items to be updated.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>