zones: Add checkDoorsHaveLocate method
When Door attributes are loaded the associated Thing also has to have a
Locate attribute so that the Door knows where it is. This is so that it
can notify observers of state changes. The only easy way of doing this
right now is to add it into the zone loader.

In the future it should be possible to add this to a PostUnmarshal
method on the attribute instead.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>