attr: Use item alias in Door.Open and Door.Close methods
The Open and Close methods currently use hardcode "OPEN DOOR" and "CLOSE
DOOR" commands when building events. However a door may not have the
alias "DOOR.

This change looks up the aliases for the Thing the Door attribute is
attached to. It then uses the first alias found to build the OPEN and
CLOSE commands dynamically for the event.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>