attr: Make door state shareable and dumpable
In order to create a fully working Door the original Door and the 'other
side' need to be able to share state. This is so that the pair will
open, close and reset seamlessly as a single door.

This change moves the shareable data into a new state struct. This is
then referenced by both sides through a pointer in the main Door struct.

The Dump method has been updated to provide the state information via a
state.dump helper.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>