event: Added jitter to event.Queue, delays in seconds
The event.Queue function now takes an additional jitter parameter.
Jitter is specified as a time.Duration and is the maximum random amount
of time that can be added to the delay period.

For an event the actual delay will be between delay and delay+jitter.

The delay and jitter durations are measured in whole seconds with the
minimum D+J being adjusted to 1 if required.

Temporally added a hardcoded jitter of 0 to attr.Door Open and Close
methods so everything compiles.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>