zones: Fix "There is no exit" for doors between zones
If a door is placed at an exit that leads between two zones the log will
show an error such as:

  There is no exit "south" for Door "the door" to block, cannot add
  other side

This is due to the fact that the zones have not been linked up yet and
so the exits between the zones have not been setup.

This change turns checkDoorsHaveOtherSide into a function (as opposed to
a method) so that it can be called after linkupZones has been called.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>