zones: Rebuilt zone loader
This commit is a rewrite of the current zone loader. When a zone is
being assembled it will now be loaded into locations (Thing with an
Exits attribute) and a temporary store (every other Thing). Each Thing
is tagged with the record from the recordjar it was created from.

The Things in the temporary store are then linked together and Things
put into temporary store inventories using references/pointers. During
the linking of Inventory now check for cyclic references.

All locations are then given an Inventory attribute and Exits are linked
together.

Finally location Inventory are populated from the temporary store using
COPIES of Things.

The temporary store is then cleared from the zone.

Once all zones are loaded Zonelinks are then linked up and the 'other
side' of doors are created.

Once all of this is complete the tagged recordjar record is dropped from
the locations.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>