attr: Have Inventory handle Locate attribute
Currently a Locate attribute is only added to a Thing if the Thing needs
to know where it is. Keeping track of what needs a Locate attribute
leads to issues. It also causes problems when something else wants to
query a Thing for its location rather than the Thing itself.

This change updates Inventory to automatically add a Locate attribute to
anything placed into an Inventory.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>