attr: Avoid copying 'other side' again
During zone loading if a Door's 'other side' is created in an
untraversed location the 'other side' will be copied again when the
location is traversed. This change adds an otherSide flag which is set
the first time the 'other side' is created. We can then check this flag
to avoid bogus copies.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>