attr,stats: Free refs properly when deleting slice element
While hunting for the reason as to why the zone temporary storage area
variables were not being garbage collected it uncovered some other
variables that were not being garbage collected either.

Currently to delete a slice element the following was used:

        s[i] = nil
        s = append(s[:i], s[i+1:])

However this only works for non-pointer/reference types. For
pointer/reference types this duplicates the last element leaving the
second copy in the backing array:

        // slice s is [*1, *2, *3, *4, *5]

        s[3] = nil

        // s is now [*1, *2, nil, *4, *5]

        s = append(s[:i], s[i+1:])

        // s is now [*1, *2, *4, *5]

However the backing array at this point is [*1, *2, *4, *5, *5] because
the last two elements are copied back by one element.

Instead the delete should be written as:

        copy(s[i:], s[i+1:])
        s[len(s)-1] = nil
        s = s[len(s)-1]

This is actually detailed in 'Slice Tricks' on the golang wiki :(

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>