stats: Another stab at calculating memory stats
Debugging memory issues has highlighted how inaccurate the current
memory value on the status log lines are.

This change changes the calculation and the values for memory usage are
now consistent with that shown by the scavenger values for consumed
memory seen when the server is run with GODEBUG=gctrace=1.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>