message: Drop Buffer interface & rename type buffer→Buffer
The Buffer interface is not very useful and is not required. Even the
justification given in the interface description "Buffer allows a buffer
to be embedded in a struct without exposing buffet itself." is bogus as
this is achieved using private fields in the buffer struct.

This change drops the Buffer interface, renames the buffer type Buffer
and updates all references for the rename.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>