message: Add reuseable *Buffer pool
Debugging memory issues has hilighted the high number of allocated and
garbage collected *message.Buffer variables. This change implements a
reusable pool of *message.Buffer. A *message.Buffer can be taken from
the pool using message.AcquireBuffer and put back into the pool using
message.ReleaseBuffer. If message.ReleaseBuffer is not called the buffer
will be discarded and the pool will replenish itself with new buffers.

For a moderate number of players (1,000) the number of allocations was
reduced by nearly 30%.

This chage only implements the pool, AcquireBuffer and ReleaseBuffer.
Another commit will change usage of &Buffer{} and NewBuffer() over to
using the pool.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>